fix(query): scope secret key authentication per project to prevent cross-project data access
SecretKeyStore now maps keys to optional project_id scopes via
HashMap<String, Option<String>>, replacing the flat Vec<String>.
check_key_for_project() rejects keys that don't match the requested
project_id, while global keys (env var, empty project_id in DB)
retain access to all projects.

All 6 route handlers now pass project_id to check_auth(). POST
handlers (funnel, retention, segment) follow parse-validate-auth
ordering for correct error codes (400 vs 401).

Includes a clippy fix: remove unnecessary as u64 cast in segment.rs

Verification:
- 40/40 unit tests pass (9 new: check_key_for_project variants)
- 32/32 integration tests pass
- 79/79 E2E tests pass (full pipeline from ingestion to query)
- clippy clean (-D warnings) for query package
- fmt clean for query package

+++ services/query/src/queries/segment.rs
@@ -46,7 +46,7 @@ LIMIT ?",
             let params = vec![
                 QueryParam::String(project_id.to_string()),
                 QueryParam::String(event_name.to_string()),
-                QueryParam::U64(within as u64),
+                QueryParam::U64(within),
                 value_to_param(&cond.value),
                 QueryParam::U64(limit),
             ];
